Fix broken systemd service file
Keep forgetting about an interesting quirk of systemd: if you explicitly
specify the user you want the service unit to be run as, you're obliged
to run it as a system-wide service, as opposed to a user-space one. This
is true even if you were to state the same user whose local systemd
folder gets used for the .service file (in our case, user pi). Funnily
enough, this still instantiates the app, but systemd considers it as
inactive and therefore doesn't collect logs from it, which is
undesirable.